CRUD operations for GradePoint

+++ M_CGPA/Model/GradePointM.cs
@@ -4,8 +4,8 @@ namespace M_CGPA.Model
     class GradePointM
     {
         public int Id { get; set; }
-        public double Minimum { get; set; }
-        public double Maximum { get; set; }
+        public int Minimum { get; set; }
+        public int Maximum { get; set; }
         public string Grade { get; set; }
         public double Point { get; set; }
     }
